Update util.js
I was struggling with an issue using socket.io client on a mobile
using the native android browser (not chrome). 
The stock browser, doesn't implement websockets, so it fallbacks to XHR.

The 100ms are fine for desktops, but on a mobile you still need to face
with an infinite spinning wheel. I've incremented to 500 and the problem
was gonne! 