Add a counter for XHR polling on IE8
IE8's `new Date()` is only accurate to 15ms[1]. This, coupled with IE8's
caching of XHR responses[2] means that IE8 can see the same message from
the server twice (it's particularly apparent with 'connect' messages).

Introduce a simple counter that, along with the timestamp, makes URLS
unique on IE8.

1. http://ejohn.org/blog/accuracy-of-javascript-time/
2. http://blogs.msdn.com/b/ieinternals/archive/2009/07/23/the-ie8-native-xmlhttprequest-object.aspx