Fix re-export of `io` module
The previous (CJS module) export semantics were of a *namespace*-style
export, while the semantics introduced in c76d367 are of a single value
exported as the default value instead. This broke downstream consumers
who were using the namespace-style export, as documented at
[socketio/socket.io#4121][4121]. While the single default export should
work fine (and might be preferable!), it's a breaking change, and
appears to be an *accidental* breaking change, so this reverts to the
previous semantics.